Implement the general search procedure

+++ src/code/Search/Heuristic.java
@@ -3,5 +3,5 @@ package code.Search;
 import code.Problem.State;
 
 public interface Heuristic {
-    double calculate(State s);
+    int calculate(State s);
 }
